+++ packages/LTable/src/LTable.vue
@@ -1,7 +1,7 @@
 <!--
  * @Date: 2020-03-26 09:43:59
  * @Author: junfeng.liu
- * @LastEditTime: 2020-07-09 18:03:56
+ * @LastEditTime: 2021-01-13 16:32:12
  * @LastEditors: junfeng.liu
  * @Description: 表格组件
 
@@ -54,7 +54,7 @@
 </template>
 
 <script>
-import { isFunction, isNull } from '@/utils/check'
+import { isFunction, isNull, isEmpty } from '@/utils/check'
 import { deepCopy } from '@/utils/util'
 
 export default {
@@ -185,11 +185,16 @@ export default {
             item.render = (h, params) => {
                 let btns = item.buttons.map((btn) => {
                     let show = true
+                    let hasPermission = true
 
                     if (isFunction(btn.show)) show = btn.show(params)
                     else if (!isNull(btn.show)) show = !!btn.show
 
-                    if (!show) return null
+                    if (isFunction(this.__l_checkPermission__) && !isEmpty(btn.permission)) {
+                        hasPermission = this.__l_checkPermission__(btn.permission)
+                    }
+
+                    if (!show || !hasPermission) return null
 
                     return (
                         <Button style={ btn.style } icon={ btn.icon } type={ btn.type || 'primary' } size={ btn.size || 'small' } onClick={ item.onClick.bind(null, (Object.assign({}, params, btn))) }>{ btn.label }</Button>
@@ -206,11 +211,16 @@ export default {
             item.render = (h, params) => {
                 let links = item.links.map((link) => {
                     let show = true
+                    let hasPermission = true
 
                     if (isFunction(link.show)) show = link.show(params)
                     else if (!isNull(link.show)) show = !!link.show
 
-                    if (!show) return null
+                    if (isFunction(this.__l_checkPermission__) && !isEmpty(link.permission)) {
+                        hasPermission = this.__l_checkPermission__(link.permission)
+                    }
+
+                    if (!show || !hasPermission) return null
 
                     return (
                         <a onClick={ item.onClick.bind(null, (Object.assign({}, link, params))) }>{ link.label }</a>